@@ -19,7 +19,7 @@
     "core-js": "^3.6.5",
     "ethereum-blockies-base64": "^1.0.2",
     "ethers": "^4.0.47",
-    "idb": "^5.0.2",
+    "idb": "^5.0.3",
     "lodash": "^4.17.15",
     "raiden-ts": "file:../raiden-ts",
     "register-service-worker": "^1.7.1",
@@ -28,12 +28,12 @@
     "typeface-roboto": "0.0.75",
     "vue": "^2.6.11",
     "vue-class-component": "^7.2.3",
-    "vue-i18n": "^8.17.5",
+    "vue-i18n": "^8.17.7",
     "vue-property-decorator": "^8.4.2",
     "vue-qrcode-reader": "^2.3.0",
-    "vue-router": "^3.1.6",
+    "vue-router": "^3.2.0",
     "vue-virtual-scroller": "^1.0.10",
-    "vuetify": "^2.2.28",
+    "vuetify": "^2.2.29",
     "vuex": "^3.4.0",
     "vuex-persist": "^2.2.0"
   },
@@ -43,13 +43,13 @@
     "@kazupon/vue-i18n-loader": "^0.5.0",
     "@mdi/font": "^5.2.45",
     "@namics/stylelint-bem": "^6.3.1",
-    "@types/jest": "^25.2.1",
-    "@types/lodash": "^4.14.150",
+    "@types/jest": "^25.2.3",
+    "@types/lodash": "^4.14.152",
     "@types/tiny-async-pool": "^1.0.0",
     "@types/vue-i18n": "^7.0.0",
     "@types/webpack": "^4.41.13",
-    "@typescript-eslint/eslint-plugin": "^2.33.0",
-    "@typescript-eslint/parser": "^2.33.0",
+    "@typescript-eslint/eslint-plugin": "^2.34.0",
+    "@typescript-eslint/parser": "^2.34.0",
     "@vue/cli": "^4.3.1",
     "@vue/cli-plugin-babel": "^4.3.1",
     "@vue/cli-plugin-eslint": "^4.3.1",
@@ -61,13 +61,13 @@
     "@vue/cli-service": "^4.3.1",
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^5.0.2",
-    "@vue/test-utils": "^1.0.2",
+    "@vue/test-utils": "^1.0.3",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.8.0",
     "eslint-import-resolver-alias": "^1.1.2",
     "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jsdoc": "^25.4.1",
+    "eslint-plugin-jsdoc": "^25.4.2",
     "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-vue": "^6.2.2",
     "eslint-plugin-vue-i18n": "^0.3.0",
@@ -79,11 +79,11 @@
     "sass": "^1.26.5",
     "sass-loader": "^8.0.2",
     "source-map-loader": "^0.2.4",
-    "stylelint": "^13.3.3",
+    "stylelint": "^13.5.0",
     "stylelint-config-recommended-scss": "^4.2.0",
     "stylelint-scss": "^3.17.2",
     "ts-jest": "25.5.1",
-    "typescript": "^3.8.3",
+    "typescript": "^3.9.3",
     "vue-cli-plugin-i18n": "^1.0.1",
     "vue-cli-plugin-vuetify": "^2.0.5",
     "vue-jest": "^3.0.5",
