@@ -1,13 +1,8 @@
-import { Observable, combineLatest } from 'rxjs';
-import { filter, scan, startWith, share, map, distinctUntilChanged } from 'rxjs/operators';
+import { Observable } from 'rxjs';
+import { filter, scan, startWith, share } from 'rxjs/operators';
 import memoize from 'lodash/memoize';
-import isEmpty from 'lodash/isEmpty';
-import isEqual from 'lodash/isEqual';
 
-import { Latest } from '../types';
 import { RaidenAction } from '../actions';
-import { Capabilities } from '../constants';
-import { RaidenConfig } from '../config';
 import { Presences, Caps } from './types';
 import { matrixPresence } from './actions';
 
@@ -67,31 +62,4 @@ export function parseCaps(caps?: string | null): Caps | undefined {
     }
     return result;
   } catch (err) {}
-}
-
-/**
- * Creates an observable which returns the actual Caps mapping, based on config and other runtime
- * conditions, such as UDC deposit balance.
- *
- * @param config$ - Observable of RaidenConfig objects
- * @param udcBalance$ - Observable of latest UDC deposit
- * @returns Observable of effective Caps mapping
- */
-export function getCaps$(
-  config$: Observable<RaidenConfig>,
-  udcBalance$: Observable<Latest['udcBalance']>,
-): Observable<Caps> {
-  return combineLatest([config$, udcBalance$]).pipe(
-    map(([{ caps, monitoringReward, rateToSvt }, udcBalance]) => ({
-      // 'noReceive' is false (i.e. receiving enabled) iff (0 < monitoringReward <= udcBalance)
-      // TODO: set per token?
-      [Capabilities.NO_RECEIVE]: !(
-        monitoringReward?.gt(0) &&
-        monitoringReward.lte(udcBalance) &&
-        !isEmpty(rateToSvt)
-      ),
-      ...caps, // default & user's config.caps has priority over keys above, but unset by default
-    })),
-    distinctUntilChanged(isEqual),
-  );
 }
