@@ -4,7 +4,6 @@ import {
   Raiden,
   RaidenPaths,
   RaidenPFS,
-  UInt,
   ErrorCodes,
   RaidenError
 } from 'raiden-ts';
@@ -13,7 +12,7 @@ import { RootState, Tokens } from '@/types';
 import { Web3Provider } from '@/services/web3-provider';
 import { BalanceUtils } from '@/utils/balance-utils';
 import { DeniedReason, Progress, Token, TokenModel } from '@/model/types';
-import { BigNumber, BigNumberish, parseEther, parseUnits } from 'ethers/utils';
+import { BigNumber, BigNumberish, parseEther } from 'ethers/utils';
 import { exhaustMap, filter } from 'rxjs/operators';
 import asyncPool from 'tiny-async-pool';
 import { ConfigProvider } from './config-provider';
@@ -42,8 +41,7 @@ export default class RaidenService {
         {
           pfsSafetyMargin: 1.1,
           pfs: process.env.VUE_APP_PFS,
-          matrixServer: process.env.VUE_APP_TRANSPORT,
-          monitoringReward: parseUnits('5', 18) as UInt<32>
+          matrixServer: process.env.VUE_APP_TRANSPORT
         },
         subkey
       );
@@ -172,6 +170,10 @@ export default class RaidenService {
         raiden.events$
           .pipe(filter(value => value.type === 'raidenShutdown'))
           .subscribe(() => this.store.commit('reset'));
+
+        raiden.config$.subscribe(config =>
+          this.store.commit('updateConfig', config)
+        );
 
         raiden.events$.subscribe(value => {
           if (value.type === 'tokenMonitored') {
