@@ -9,6 +9,7 @@ import store from '@/store/index';
 import Vuetify from 'vuetify';
 import RaidenService from '@/services/raiden-service';
 import App from '@/App.vue';
+import { Capabilities } from 'raiden-ts';
 
 Vue.use(VueRouter);
 Vue.use(Vuex);
@@ -30,7 +31,7 @@ describe('App.vue', () => {
     wrapper = shallowMount(App, {
       vuetify,
       store,
-      stubs: ['router-view'],
+      stubs: ['router-view', 'v-dialog'],
       mocks: {
         $router: router,
         $raiden: $raiden,
@@ -49,5 +50,21 @@ describe('App.vue', () => {
     wrapper.vm.$destroy();
 
     expect($raiden.disconnect).toHaveBeenCalledTimes(1);
+  });
+
+  test("show ReceivingDiabled dialog if can't receive", async () => {
+    expect.assertions(2);
+
+    store.commit('updateConfig', {
+      caps: { [Capabilities.NO_RECEIVE]: false }
+    });
+    await wrapper.vm.$nextTick();
+    expect(wrapper.vm.$data.showReceivingDisabled).toBe(false);
+
+    store.commit('updateConfig', {
+      caps: { [Capabilities.NO_RECEIVE]: true }
+    });
+    await wrapper.vm.$nextTick();
+    expect(wrapper.vm.$data.showReceivingDisabled).toBe(true);
   });
 });
