@@ -2,11 +2,11 @@
 export { Raiden } from './raiden';
 export { RaidenState, encodeRaidenState } from './state';
 export { RaidenEvent, RaidenAction } from './actions';
-export { ShutdownReason } from './constants';
 export { RaidenTransfer, RaidenTransferStatus } from './transfers/state';
 export { ChannelState, RaidenChannel, RaidenChannels } from './channels/state';
 export { RaidenPaths, RaidenPFS } from './services/types';
 export { RaidenConfig } from './config';
+export * from './constants';
 export * from './types';
 export * from './utils/types';
 export * from './utils/error';
